requirements.txt
discord.py[voice]==2.5.2
yt-dlp==2024.12.13
PyNaCl==1.5.0
python-dotenv==1.0.1
requests==2.32.3
ffmpeg-python==0.2.0
psutil==5.9.8
